chore: refactor registration page, add registration form
