+++ packages/client/src/pages/registration/Registration.tsx
@@ -1,92 +1,19 @@
-import { Link } from 'react-router-dom'
-
-import { Anchor, Box, Button, TextInput, Title } from '@mantine/core'
-import { useForm } from '@mantine/form'
+import { Title } from '@mantine/core'
 
 import { FormWrapper } from '@/components'
-import { useRoutes } from '@/hooks'
-import {
-  firstNameRule,
-  loginRule,
-  passwordRule,
-  phoneRule,
-  secondNameRule,
-} from '@/shared'
+import { RegistrationForm } from '@/pages/registration/components'
 
 export const Registration = (): JSX.Element => {
-  const { paths } = useRoutes()
-  const form = useForm({
-    validateInputOnBlur: true,
-    initialValues: {
-      first_name: '',
-      second_name: '',
-      login: '',
-      email: '',
-      password: '',
-      phone: '',
-    },
-
-    validate: {
-      first_name: value => firstNameRule(value),
-      second_name: value => secondNameRule(value),
-      login: value => loginRule(value),
-      email: value => firstNameRule(value),
-      password: value => passwordRule(value),
-      phone: value => phoneRule(value),
-    },
-  })
-
   return (
     <FormWrapper formId="registration">
       <Title align="center" mb={16}>
         Регистрация
       </Title>
-      <form onSubmit={form.onSubmit(values => console.log(values))}>
-        <TextInput
-          withAsterisk
-          mb={32}
-          label="Имя"
-          {...form.getInputProps('first_name')}
-        />
-        <TextInput
-          withAsterisk
-          mb={32}
-          label="Фамилия"
-          {...form.getInputProps('second_name')}
-        />
-        <TextInput
-          withAsterisk
-          mb={32}
-          label="Логин"
-          {...form.getInputProps('login')}
-        />
-        <TextInput
-          withAsterisk
-          mb={32}
-          label="Почта"
-          {...form.getInputProps('email')}
-        />
-        <TextInput
-          withAsterisk
-          mb={32}
-          label="Пароль"
-          {...form.getInputProps('password')}
-        />
-        <TextInput
-          withAsterisk
-          mb={32}
-          label="Телефон"
-          {...form.getInputProps('phone')}
-        />
-        <Button mb={16} fullWidth={true} type="submit">
-          Зарегистрироваться
-        </Button>
-        <Box ta="center">
-          <Anchor component={Link} to={paths.Login}>
-            Уже зарегистрированы?
-          </Anchor>
-        </Box>
-      </form>
+      <RegistrationForm
+        handleSubmit={data => {
+          console.log(data)
+        }}
+      />
     </FormWrapper>
   )
 }
