+++ packages/client/src/pages/registration/components/index.ts
@@ -1,0 +1,1 @@
+export * from './registration-form'
