+++ packages/client/src/pages/registration/components/registration-form/index.ts
@@ -1,0 +1,1 @@
+export * from './RegistrationForm'
